restructured project, added clean.sh, build.sh
