+++ .gitignore
@@ -1,10 +1,12 @@
-.metadata
-.recommenders
-bin
-.DS_Store
 .classpath
+.DS_Store
+.metadata
 .project
-build
+.recommenders
 .settings
-output/spark-final.conf
+build
+bin/build
+bin/sparkaid.jar
+eclipse-bin
 output/spark.conf.advise
+output/spark-final.conf
