+++ auto-config/core/src/main/common/Common.java
@@ -1,7 +1,7 @@
 package core.src.main.common;
 import java.util.*;
 
-import utils.UtilsConversion;
+import utils.src.main.UtilsConversion;
 
 public class Common {
