+++ bin/build.txt
@@ -1,0 +1,6 @@
+../auto-config/console/src/main/ConfigurationConsole.java
+../auto-config/core/src/main/common/Common.java
+../auto-config/core/src/main/yarn/Yarn.java
+../auto-config/dynamicallocation/src/main/DynamicAllocation.java
+../auto-config/streaming/src/main/Streaming.java
+../utils/src/main/UtilsConversion.java
