+++ bin/clean.sh
@@ -1,0 +1,9 @@
+rm -r build
+rm sparkaid.jar
+rm ../output/*
+
+rm -r ../auto-config/console/bin		
+rm -r ../auto-config/dynamicallocation/bin	
+rm -r ../auto-config/core/bin		
+rm -r ../auto-config/streaming/bin	
+rm -r ../utils/bin
