+++ bin/run.sh
@@ -1,0 +1,3 @@
+./build.sh
+
+java -classpath sparkaid.jar console.src.main.ConfigurationConsole "$@"
